moved grid 1 dsturcture functions
