+++ autoarray/structures/grids/uniform_1d.py
@@ -325,57 +325,3 @@ class Grid1D(Structure):
         )
 
         return Grid2DIrregular(values=grid)
-
-    def structure_2d_from(
-        self, result: np.ndarray
-    ) -> Union[Array1D, Grid2D, Grid2DTransformed, Grid2DTransformedNumpy]:
-        """
-        Convert a result from an ndarray to an aa.Array2D or aa.Grid2D structure, where the conversion depends on
-        type(result) as follows:
-
-        .. code-block:: bash
-
-            - 1D np.ndarray   -> aa.Array2D
-            - 2D np.ndarray   -> aa.Grid2D
-
-        This function is used by the grid_2d_to_structure decorator to convert the output result of a function
-        to an autoarray structure when a `Grid2D` instance is passed to the decorated function.
-
-        Parameters
-        ----------
-        result
-            The input result (e.g. of a decorated function) that is converted to a PyAutoArray structure.
-        """
-        from autoarray.structures.arrays.uniform_1d import Array1D
-        from autoarray.structures.grids.transformed_2d import Grid2DTransformed
-        from autoarray.structures.grids.transformed_2d import Grid2DTransformedNumpy
-        from autoarray.structures.grids.uniform_2d import Grid2D
-
-        if len(result.shape) == 1:
-            return Array1D(values=result, mask=self.mask)
-
-        if isinstance(result, Grid2DTransformedNumpy):
-            return Grid2DTransformed(values=result, mask=self.mask)
-        return Grid2D(values=result, mask=self.mask.derive_mask.to_mask_2d)
-
-    def structure_2d_list_from(
-        self, result_list: List
-    ) -> List[Union[Array1D, Grid2D, Grid2DTransformed, Grid2DTransformedNumpy]]:
-        """
-        Convert a result from a list of ndarrays to a list of aa.Array2D or aa.Grid2D structure, where the conversion
-        depends on type(result) as follows:
-
-        .. code-block:: bash
-
-            - [1D np.ndarray] -> [aa.Array2D]
-            - [2D np.ndarray] -> [aa.Grid2D]
-
-        This function is used by the grid_like_list_to_structure-list decorator to convert the output result of a
-        function to a list of autoarray structure when a `Grid2D` instance is passed to the decorated function.
-
-        Parameters
-        ----------
-        result_list
-            The input result (e.g. of a decorated function) that is converted to a PyAutoArray structure.
-        """
-        return [self.structure_2d_from(result=result) for result in result_list]
